3.2-crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
pip install django-filter